requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.0
openai>=1.12.0
python-docx>=1.1.0
PyPDF2>=3.0.1
python-multipart>=0.0.6
numpy>=1.26.4
pandas>=2.2.1
scikit-learn>=1.4.1
GitPython>=3.1.42
langchain>=0.1.11
langchain-openai>=0.0.8
docx2txt>=0.8
